[move file] move components to organisms
