+++ src/App.jsx
@@ -2,8 +2,7 @@ import React from "react"
 import IncidentsOfTheMonth from "./components/organisms/incidentsOfTheMonth/incidentsOfTheMonth"
 import CommingIssue from './components/organisms/commingIssue/commingIssue'
 import EstablishmentInfo from './components/organisms/establishmentInfo/establishmentInfo'
-import CommingIssue from './components/organismes/commingIssue/commingIssue'
-import { DropdownContainer } from './components/toGoToOrganisms/DropdownContainer'
+import { DropdownContainer } from './components/organisms/DropDownIncidentsTypes/DropdownContainer'
 
 
 function App() {
