+++ src/components/organisms/DropDownIncidentsTypes/dropdownContainer.scss
