+++ src/styles/main.scss
@@ -18,7 +18,6 @@
 @import '../components/atoms/variationStat/variationStat.scss';
 @import '../components/atoms/mainStat/mainStat.scss';
 @import '../components/atoms/tagsIncidents/tagsIncidents.scss';
-@import '../components/atoms/dateIncidents/dateIncidents.scss';
 @import '../components/atoms/localisationIncidents/localisationnIncidents.scss';
 
 // Molecules
@@ -31,7 +30,6 @@
 @import '../components/organisms/incidentsOfTheMonth/incidentsOfTheMonth.scss';
 @import '../components/organisms/establishmentInfo/establishmentInfo.scss';
 @import '../components/organisms/DropDownIncidentsTypes/dropdownIncidentsTypese.scss';
-@import "../components/organismes/DropDownIncidentsTypes/dropdownContainer.scss";
 
 // Templates
 @import '../components/templates/card.scss';
